requirement.txt
pyupbit
pandas
numpy
python-dotenv
requests
slack-sdk